Delete the typechecking debug outputs from a long time ago
Dusk used to be able to generate GraphViz graphs of all the nodes in the
program, then launch them in a webbrowser. This was usually very slow, and there many shortcomings with the quality of the final output. DVD has the potential to be way, way better and more useful.

Dusk also used to be able to tell you how the constraints changed from one level to the next, per pass, per unit. The signal to noise ratio was too low to be useful. DVD once again has the potential to fill the same role much better.