Add mir instruction for a new type of intrinsic
The idea is I will have a macro which lets me define Dusk functions and
methods as ordinary Rust code. The macro will do the work of registering
the functions/methods in all the proper places, as well as generating
thunks which convert between Rust and Dusk values in parameters and
return values.