Use message passing to call userspace functions
This provides a foundation for moving the interpreter out-of-process.

Step 1: Split `DvmServerCoordinator` into `DvmServerCoordinator` and
`DvmClientCoordinator`, with the two sending serialized messages and
responses back and forth.
Step 2: Replace all calls to `Driver` and `DriverRef`-dependent
functions with calls to `dvm::send_message()`
Step 3: Move DVM server into its own process and connect to it the same
way I do with DVD.