Refactor
Use dyn TypeProviders everywhere, and start paving the way for
typechecking SUHMM dependees (aka, for all intents and purposes,
extend blocks)

The basic idea is, for each SUHMM dependee (extend block), we create a
dedicated Levels/Units instance, along with some graph state as well. We
then attempt to fully typecheck each dependee. If we fail, we discard
everything we just did and move on to the next dependee. If we succeed,
we save everything we just did in the RealTypeProvider and move on. We
continue to retry each remaining un-typechecked SUHMM dep until either
all fail to typecheck, or there are none (i.e. they all typechecked
successfully).