Move slow, and somehow still break things
This commit tries to only typecheck the dependencies of SUHMM dependees while in SUHMM mode (for lack of a better, but still succinct description). But currently it's failing on the assertion on line 590. On further investigation, all outstanding components are being ruled out by `remove_comps_with_outstanding_deps`, and are actually from unrelated pieces of code in the core library.