Delete DVD
The main motivation for this is getting rid of some (or all) of the github security advisories. But also, I don't think DVD really achieved what I wanted it to anyway. If I revisit the concept in the future, I think I would ditch the idea of anticipating what kind of information from the compiler that DVD might want and sending it via IPC, and instead produce visualizations directly based on compiler data structures.