zip: Implement central directory-aligned files
Motivation: I noticed when running apksigner that the files used for signing an APK with the v1 scheme (aka JAR signing) are placed right before the APK signature block. I don't know whether this is required, or simply an optimization so that the v1 and v2+ signatures are closer to each other in the APK. Regardless, by allowing a separate list of files to be aligned to the central directory (which is immediately displaced by the APK signature block), I can replicate this behaviour, if I ever end up implementing v1 signatures.