destroy everything i guess
Was working on adding bindings to decision tree, then got sidetracked with wanting to not duplicate data between pattern matching "paths" already generated during parsing, and created a parse-time "pattern matching context" that contains scrutinee value ids, then creating an updated version of this data structure at typechecking time. The problem with this approach is that enum patterns won't necessarily have payload patterns, even if there is a payload type. But we need the payload pattern for its scrutinee value and type for constructing the decision tree.